--- aequilibrae/log/__init__.py
@@ -1,1 +1,0 @@
-from .log import Log
